# lab_3/products/openapi.json
{"openapi": "3.1.0", "info": {"title": "Product Catalog API", "description": "API \u0434\u043b\u044f \u0443\u043f\u0440\u0430\u0432\u043b\u0435\u043d\u0438\u044f \u043a\u0430\u0442\u0430\u043b\u043e\u0433\u043e\u043c \u0442\u043e\u0432\u0430\u0440\u043e\u0432", "version": "1.0.0"}, "paths": {"/products/list": {"get": {"tags": ["products"], "summary": "Get Products", "operationId": "get_products_products_list_get", "parameters": [{"name": "name", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "\u0424\u0438\u043b\u044c\u0442\u0440 \u043f\u043e \u043d\u0430\u0437\u0432\u0430\u043d\u0438\u044e \u0442\u043e\u0432\u0430\u0440\u0430", "title": "Name"}, "description": "\u0424\u0438\u043b\u044c\u0442\u0440 \u043f\u043e \u043d\u0430\u0437\u0432\u0430\u043d\u0438\u044e \u0442\u043e\u0432\u0430\u0440\u0430"}, {"name": "category", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "\u0424\u0438\u043b\u044c\u0442\u0440 \u043f\u043e \u043a\u0430\u0442\u0435\u0433\u043e\u0440\u0438\u0438", "title": "Category"}, "description": "\u0424\u0438\u043b\u044c\u0442\u0440 \u043f\u043e \u043a\u0430\u0442\u0435\u0433\u043e\u0440\u0438\u0438"}, {"name": "min_price", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "\u041c\u0438\u043d\u0438\u043c\u0430\u043b\u044c\u043d\u0430\u044f \u0446\u0435\u043d\u0430", "title": "Min Price"}, "description": "\u041c\u0438\u043d\u0438\u043c\u0430\u043b\u044c\u043d\u0430\u044f \u0446\u0435\u043d\u0430"}, {"name": "max_price", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "\u041c\u0430\u043a\u0441\u0438\u043c\u0430\u043b\u044c\u043d\u0430\u044f \u0446\u0435\u043d\u0430", "title": "Max Price"}, "description": "\u041c\u0430\u043a\u0441\u0438\u043c\u0430\u043b\u044c\u043d\u0430\u044f \u0446\u0435\u043d\u0430"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Product"}, "title": "Response Get Products Products List Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/products/details": {"get": {"tags": ["products"], "summary": "Get Product", "operationId": "get_product_products_details_get", "parameters": [{"name": "product_id", "in": "query", "required": true, "schema": {"type": "integer", "title": "Product Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Product"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/products/create": {"post": {"tags": ["products"], "summary": "Create Product", "operationId": "create_product_products_create_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Product"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/products/update": {"put": {"tags": ["products"], "summary": "Update Product", "operationId": "update_product_products_update_put", "parameters": [{"name": "product_id", "in": "query", "required": true, "schema": {"type": "integer", "title": "Product Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductBase"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Product"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/products/delete": {"delete": {"tags": ["products"], "summary": "Delete Product", "operationId": "delete_product_products_delete_delete", "parameters": [{"name": "product_id", "in": "query", "required": true, "schema": {"type": "integer", "title": "Product Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Product"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/products/create/bulk": {"post": {"tags": ["products"], "summary": "Create Products Bulk", "operationId": "create_products_bulk_products_create_bulk_post", "requestBody": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ProductCreate"}, "type": "array", "title": "Products"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Product"}, "type": "array", "title": "Response Create Products Bulk Products Create Bulk Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"tags": ["root"], "summary": "Read Root", "operationId": "read_root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/openapi.json": {"get": {"tags": ["root"], "summary": "Get Operapi", "operationId": "get_operapi_openapi_json_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Product": {"properties": {"name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "price": {"type": "string", "title": "Price"}, "category": {"type": "string", "title": "Category"}, "in_stock": {"type": "integer", "title": "In Stock"}, "id": {"type": "integer", "title": "Id"}}, "type": "object", "required": ["name", "description", "price", "category", "in_stock", "id"], "title": "Product"}, "ProductBase": {"properties": {"name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "price": {"anyOf": [{"type": "number"}, {"type": "string"}], "title": "Price"}, "category": {"type": "string", "title": "Category"}, "in_stock": {"type": "integer", "title": "In Stock"}}, "type": "object", "required": ["name", "description", "price", "category", "in_stock"], "title": "ProductBase"}, "ProductCreate": {"properties": {"name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "price": {"anyOf": [{"type": "number"}, {"type": "string"}], "title": "Price"}, "category": {"type": "string", "title": "Category"}, "in_stock": {"type": "integer", "title": "In Stock"}}, "type": "object", "required": ["name", "description", "price", "category", "in_stock"], "title": "ProductCreate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}